pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "medical-graphrag-factchecker"
version = "1.0.0"
description = "AI-powered multilingual medical fact-checking system using GraphRAG and Knowledge Graphs"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
authors = [
    {name = "Sourav Kumar Sharma", email = "sourav@example.com"}
]
keywords = ["medical", "fact-checking", "graphrag", "knowledge-graph", "neo4j", "rag", "ai", "healthcare"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "datasets>=2.19.0",
    "beautifulsoup4>=4.12.3",
    "requests>=2.31.0",
    "sentence-transformers>=3.0.1",
    "numpy>=1.26.4",
    "neo4j>=5.20.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "isort>=5.13.2",
    "flake8>=7.1.1",
    "mypy>=1.13.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
